fix warning

+++ nvme/scheduler/graph_state.h
@@ -87,7 +87,7 @@ struct GraphState {
   }
 
   void SyncCurrentRoundPending() {
-    for (int i = 0; i < num_blocks_; i++) {
+    for (size_t i = 0; i < num_blocks_; i++) {
       current_round_pending_.at(i) = true;
     }
   }
